Rework to remove master/slave notion and orchestrate bootstrap
This is a major overhaul of the formula; the entire notion of
master/slave is non-existent in galera so there should not be such a
distinction in the pillar data and states. The normal node
installation will simply try to join an existing cluster; the special
case of installing a new cluster is handled by orchestration.